backend/requirements.txt
Flask==3.0.0
instagrapi==2.1.2
python-dotenv==1.0.0
Pillow>=8.1.1
